refactor: Move branches code into a mixin
