--- guide/.vuepress/branches.js
@@ -1,12 +1,0 @@
-export default [
-	{
-		label: 'v12 (stable)',
-		version: '12.x',
-		aliases: ['12', 'stable'],
-	},
-	{
-		label: 'v11',
-		version: '11.x',
-		aliases: ['11'],
-	},
-];
